Removed unused variables, gave a switch statement a missing default clause and made an explicit cast to avoid a comparison between a signed integer expression and an unsigned integer expression

+++ castor/vdqm/handler/TapeDriveConsistencyChecker.cpp
@@ -315,7 +315,7 @@ void castor::vdqm::handler::TapeDriveConsistencyChecker::checkAssignConsistency(
 				throw ex;
       }
       else {
-      	if ( (tapeRequest->id() != ptr_driveRequest->VolReqID)) {
+      	if ( ((int)tapeRequest->id() != ptr_driveRequest->VolReqID)) {
       			
 	 				castor::exception::Exception ex(EVQBADID);
     			ex.getMessage() << "TapeDriveConsistencyChecker::checkAssignConsistency(): "
